fix: fix up address service logic
