+++ src/main/java/com/find/doongji/address/loader/AddressMappingLoader.java
@@ -59,6 +59,10 @@ public class AddressMappingLoader implements CommandLineRunner {
 
                 List<AptInfo> aptInfoList = aptRepository.selectAptInfoByUmdNmAndJibun(umdNm, jibun);
                 for (AptInfo aptInfo : aptInfoList) {
+                    String aptInfoRoadAddress = aptInfo.getRoadNm() + " " + aptInfo.getRoadNmBonbun();
+                    if (!aptInfo.getRoadNmBubun().equals("0")) {
+                        aptInfoRoadAddress += "-" + aptInfo.getRoadNmBubun();
+                    }
                     AddressMapping addressMapping = AddressMapping.builder()
                             .oldAddress(oldAddress)
                             .umdNm(umdNm)
@@ -69,7 +73,7 @@ public class AddressMappingLoader implements CommandLineRunner {
                             .roadNm(aptInfo.getRoadNm())
                             .roadNmBonbun(aptInfo.getRoadNmBonbun())
                             .roadNmBubun(aptInfo.getRoadNmBubun())
-                            .roadAddress(roadAddress)
+                            .roadAddress(aptInfoRoadAddress)
                             .aptNm(aptInfo.getAptNm())
                             .build();
                     mappings.add(addressMapping);
