+++ src/main/java/com/find/doongji/address/service/AddressService.java
@@ -1,0 +1,6 @@
+package com.find.doongji.address.service;
+
+public interface AddressService {
+
+    Long createAddressMapping(String jibunAddress, String roadAddress) throws Exception;
+}
