+++ src/main/java/com/find/doongji/danji/service/BasicDanjiService.java
@@ -1,11 +1,13 @@
 package com.find.doongji.danji.service;
 
+import com.find.doongji.address.util.AddressUtil;
 import com.find.doongji.danji.client.DanjiClient;
 import com.find.doongji.danji.payload.request.DanjiEntity;
 import com.find.doongji.danji.repository.DanjiRepository;
 import com.find.doongji.location.repository.LocationRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -18,6 +20,7 @@ public class BasicDanjiService implements DanjiService {
     private final DanjiClient danjiClient;
 
     @Override
+    @Transactional
     public void loadDanji() throws Exception {
         if (danjiRepository.checkIfTableExists() != 0) {
             System.out.println("Danji table already exists");
@@ -31,7 +34,7 @@ public class BasicDanjiService implements DanjiService {
 
         List<DanjiEntity> entities = responses.stream()
                 .map(item -> DanjiEntity.builder()
-                        .as1(item.getAs1())
+                        .as1(AddressUtil.cleanAddress(item.getAs1()))
                         .as2(item.getAs2())
                         .as3(item.getAs3())
                         .bjdCode(item.getBjdCode())
@@ -39,6 +42,8 @@ public class BasicDanjiService implements DanjiService {
                         .kaptName(item.getKaptName())
                         .build())
                 .toList();
+
+        danjiRepository.bulkInsertDanji(entities);
 
         System.out.println("End loading danji table: " + System.currentTimeMillis());
     }
