+++ src/main/java/com/find/doongji/listing/service/BasicListingService.java
@@ -1,13 +1,11 @@
 package com.find.doongji.listing.service;
 
-import com.find.doongji.address.payload.request.AddressMapping;
 import com.find.doongji.address.payload.response.AddressMappingResponse;
 import com.find.doongji.address.repository.AddressRepository;
+import com.find.doongji.address.service.AddressService;
 import com.find.doongji.address.util.AddressUtil;
 import com.find.doongji.apt.client.AptClient;
-import com.find.doongji.apt.payload.response.AptInfo;
 import com.find.doongji.apt.repository.AptRepository;
-import com.find.doongji.danji.payload.response.DanjiCode;
 import com.find.doongji.danji.repository.DanjiRepository;
 import com.find.doongji.listing.payload.request.ListingCreateRequest;
 import com.find.doongji.listing.payload.request.ListingEntity;
@@ -37,6 +35,8 @@ public class BasicListingService implements ListingService {
     private final AddressRepository addressRepository;
     private final LocationRepository locationRepository;
 
+    private final AddressService addressService;
+
     private final AptClient aptClient;
     private final DanjiRepository danjiRepository;
 
@@ -46,11 +46,11 @@ public class BasicListingService implements ListingService {
     @Override
     public void addListing(ListingCreateRequest request, MultipartFile image) throws Exception {
         String username = SecurityContextHolder.getContext().getAuthentication().getName();
-        List<AddressMappingResponse> addressMappings = addressRepository.selectAddressMappingByRoadAddress(request.getRoadAddress());
+        List<AddressMappingResponse> addressMappings = addressRepository.selectAddressMappingByJibunAddress(request.getJibunAddress());
 
         List<Long> danjiIds = new ArrayList<>();
         if (addressMappings.isEmpty()) {
-            danjiIds.add(createAddressMapping(request));
+            danjiIds.add(addressService.createAddressMapping(request.getJibunAddress(), request.getRoadAddress()));
         } else {
             for (AddressMappingResponse addressMapping : addressMappings) {
                 danjiIds.add(addressMapping.getDanjiId());
@@ -152,55 +152,6 @@ public class BasicListingService implements ListingService {
     public String removeUploadsPrefix(String path) {
         if (path == null) return null;
         return path.replaceFirst("^\\.\\/uploads\\\\?", "");
-    }
-
-    private String getDongCode(String oldAddress) {
-        String[] parts = oldAddress.split("\\s+");
-        if (parts.length < 3) {
-            throw new IllegalArgumentException("지번주소 형식이 잘못되었습니다.");
-        }
-        String siDo = parts[0];           // 시도
-        String siGunGu = parts[1];       // 시군구
-        String eupMyeonDongRi = parts[2]; // 읍면동/리
-        return locationRepository.selectDongCode(siDo, siGunGu, eupMyeonDongRi);
-    }
-
-    private Long createAddressMapping(ListingCreateRequest request) throws Exception {
-        String bjdCode = getDongCode(request.getJibunAddress());
-        List<DanjiCode> danjiCodes = danjiRepository.selectAllByBjdCode(bjdCode);
-
-        AddressUtil.AddressComponents components = AddressUtil.parseAddress(request.getRoadAddress());
-
-        // TODO: 해당 레포지토리 메서드는 단일 레코드가 아닌 집합을 리턴하기 때문에 모두 고려해야 한다
-        List<AptInfo> aptInfoList = aptRepository.selectAptInfoByRoadComponents(components.getRoadNm(), components.getRoadNmBonbun(), components.getRoadNmBubun());
-        if (aptInfoList.isEmpty()) {
-            throw new Exception("해당 도로명주소에 대한 아파트 정보가 존재하지 않습니다.");
-        }
-        AptInfo aptInfo = aptInfoList.get(0);
-        Long danjiId = null;
-        for (DanjiCode danjiCode : danjiCodes) {
-            danjiId = (aptInfo.getAptSeq() + danjiCode.getKaptCode()).hashCode() & 0xffffffffL;
-            if (!AddressUtil.cleanAddress(request.getJibunAddress()).startsWith(AddressUtil.cleanAddress(danjiCode.getSiGugunDong())))
-                continue;
-            AddressUtil.OldAddressComponents oldAddressComponents = AddressUtil.parseOldAddress(request.getJibunAddress());
-            addressRepository.insertAddressMapping(
-                    AddressMapping.builder()
-                            .roadAddress(AddressUtil.cleanAddress(request.getRoadAddress()))
-                            .oldAddress(oldAddressComponents.getJibunAddress())
-                            .umdNm(aptInfo.getUmdNm())
-                            .jibun(aptInfo.getJibun())
-                            .aptSeq(aptInfo.getAptSeq())
-                            .bjdCode(danjiCode.getBjdCode())
-                            .roadNm(components.getRoadNm())
-                            .roadNmBonbun(components.getRoadNmBonbun())
-                            .roadNmBubun(components.getRoadNmBubun())
-                            .danjiId(danjiId)
-                            .aptNm(oldAddressComponents.getAptName())
-                            .build()
-            );
-            break;
-        }
-        return danjiId;
     }
 
     private boolean checkEditPermission(Long id) {
