+++ src/main/resources/db/mapper/address.xml
@@ -12,7 +12,6 @@
         END;
     </select>
 
-
     <insert id="bulkInsertAddressMapping" parameterType="list">
         INSERT INTO address_mapping (old_address, road_address, umd_nm, jibun, apt_seq, danji_id, bjd_code, apt_nm,
         road_nm, road_nm_bonbun, road_nm_bubun)
@@ -46,6 +45,12 @@
             resultType="com.find.doongji.address.payload.response.AddressMappingResponse">
         SELECT * FROM address_mapping
         WHERE road_address = #{roadAddress}
+    </select>
+
+    <select id="selectAddressMappingByJibunAddress" parameterType="String"
+            resultType="com.find.doongji.address.payload.response.AddressMappingResponse">
+        SELECT * FROM address_mapping
+        WHERE jibun_address like CONCAT('%', #{jibunAddress}, '%')
     </select>
 
     <select id="selectAddressMappingByDanjiIdList" parameterType="list" resultType="com.find.doongji.address.payload.response.AddressMappingResponse">
